backend_wordcloud_py/pyproject.toml
[project]
name = ""
version = ""
description = ""
authors = [
    {name = "pcsishun", email = "pcsishun@gmail.com"},
]
dependencies = [
    "google-cloud-datastore>=2.11.1",
    "requests>=2.28.1",
    "flask>=2.2.2",
    "python-dotenv>=0.21.0",
    "matplotlib>=3.6.2",
    "wordcloud>=1.8.2.2",
    "Flask-Cors>=3.0.10",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[build-system]
requires = ["pdm-pep517>=1.0.0"]
build-backend = "pdm.pep517.api"

[tool]
[tool.pdm]
